added countdown service , adjusted CountDownFragment.kt

+++ app/src/main/java/com/lexneoapps/motivodoroapp/data/cdtimer/CDTimer.kt
@@ -6,10 +6,11 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "cdtimer_table")
 data class CDTimer(
-        val countdown: Long = 0L,
-        val pause: Long = 0L,
+        val countdown: Int = 0,
+        val pause: Int = 0,
         val repeat: Int = 1,
         val name: String = "",
+//        val color: Int = -30800,
         @PrimaryKey(autoGenerate = true) val id: Int = 0
 ) {
 }
